full update

+++ drivers/char/xilinx_hwicap/xilinx_hwicap.c
@@ -624,7 +624,7 @@ static int __devinit hwicap_setup(struct device *dev, int id,
 	if (!request_mem_region(drvdata->mem_start,
 					drvdata->mem_size, DRIVER_NAME)) {
 		dev_err(dev, "Couldn't lock memory region at %Lx\n",
-			regs_res->start);
+			(unsigned long long)regs_res->start);
 		retval = -EBUSY;
 		goto failed1;
 	}
@@ -645,7 +645,7 @@ static int __devinit hwicap_setup(struct device *dev, int id,
 
 	dev_info(dev, "ioremap %lx to %p with size %Lx\n",
 		 (unsigned long int)drvdata->mem_start,
-			drvdata->base_address, drvdata->mem_size);
+			drvdata->base_address, (unsigned long long)drvdata->mem_size);
 
 	cdev_init(&drvdata->cdev, &hwicap_fops);
 	drvdata->cdev.owner = THIS_MODULE;
